requirements.txt
future==0.18.2
iso8601==0.1.12
numpy==1.19.1
pandas==1.1.1
psycopg2==2.8.5
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
RPi.GPIO==0.7.0
serial==0.0.97
six==1.15.0
SQLAlchemy==1.3.19
